updated main script with server script

+++ make_setup_test.py
@@ -77,13 +77,21 @@ def exitWithMessageIfError(stream, process, errString):
 		print('Error received:')
 		print(stream.decode("utf-8"))
 		if process is not None:
-			process.terminate()
+			if isinstance(process, list):
+				for proc in process:
+					proc.terminate()
+			else:
+				process.terminate()
 		sys.exit(errString)
 
 
 def debugPrint(string):
 	if debug == 1:
 		print(string)
+
+def debugPrintNNewLine(string):
+	if debug == 1:
+		print(string, end='', flush=True)
 
 confDefault = [
 	'regtest=1',
@@ -93,8 +101,7 @@ confDefault = [
 	'listen=1',
 	'dbcache=50',
 	'whitelist=127.0.0.1',
-	'node_dir_placeholder',
-	'load_block_placeholder'
+	'node_dir_placeholder'
 ]
 
 confRegtest = [
@@ -130,94 +137,58 @@ for node in range(0, num_clients):
 
 
 ####### create starting blockchain
-print('creating starting block')
-# create directory
-os.chdir(nodesPath)
-if 'starting_chain_node' in os.listdir():
-	shutil.rmtree('starting_chain_node')
-os.mkdir('starting_chain_node')
+# print('creating starting block')
+# # create directory
+# os.chdir(nodesPath)
+# if 'starting_chain_node' in os.listdir():
+# 	shutil.rmtree('starting_chain_node')
+# os.mkdir('starting_chain_node')
 
 # create and make conf file
-os.chdir(binPath)
-node = -1
-confThisNode = confDefault.copy()
-confThisNodeRegTest = confRegtest.copy()
-nodeDir = nodesPath + "starting_chain_node"
-confThisNode[7] = 'datadir=' + nodeDir
-confThisNode[8] = 'loadblock=' + nodeDir + 'regtest/blocks/blk00000.dat'
-port = BASE_PORT_NUM + node
-confThisNodeRegTest[0] = 'port=' + str(port)
-rpcport = BASE_RPC_PORT_NUM + node
-confThisNodeRegTest[1] = 'rpcport=' + str(rpcport)
-confFilePath = nodeDir + delim + "bitcoin.conf"
-contentDefault = '\n'.join(confThisNode)
-contentRegTest = '\n'.join(confThisNodeRegTest)
-with open(confFilePath, "w") as text_file:
-	text_file.write(contentDefault)
-	text_file.write('\n\n[regtest]\n')
-	text_file.write(contentRegTest)
+# os.chdir(binPath)
+# node = -1
+# confThisNode = confDefault.copy()
+# confThisNodeRegTest = confRegtest.copy()
+# nodeDir = nodesPath + "starting_chain_node"
+# confThisNode[7] = 'datadir=' + nodeDir
+# confThisNode[8] = 'loadblock=' + nodeDir + 'regtest/blocks/blk00000.dat'
+# port = BASE_PORT_NUM + node
+# confThisNodeRegTest[0] = 'port=' + str(port)
+# rpcport = BASE_RPC_PORT_NUM + node
+# confThisNodeRegTest[1] = 'rpcport=' + str(rpcport)
+# confFilePath = nodeDir + delim + "bitcoin.conf"
+# contentDefault = '\n'.join(confThisNode)
+# contentRegTest = '\n'.join(confThisNodeRegTest)
+# with open(confFilePath, "w") as text_file:
+# 	text_file.write(contentDefault)
+# 	text_file.write('\n\n[regtest]\n')
+# 	text_file.write(contentRegTest)
 
 #run bitcoind
-bitcoindCmdArgs[1] = '-conf=' + confFilePath
-btcStartingChain = subprocess.Popen(bitcoindCmdArgs, stdout=subprocess.DEVNULL, stderr=subprocess.STDOUT)
-debugPrint("	started bitcoind of starting chain node")
-
-debugPrint("	waiting for client to finish setup...")
-time.sleep(5)
-initCmdArgs = cliCmdArgs.copy()
-initCmdArgs[3] = '-rpcport=' + str(rpcport)
-
-# get address for Txs
-GetAddressCmdArgs = initCmdArgs.copy()
-GetAddressCmdArgs.append('getnewaddress')
-getAddrRet = subprocess.run(GetAddressCmdArgs, capture_output=True)
-exitWithMessageIfError(getAddrRet.stderr, btcStartingChain, 'Error getting address')
-addressStr = getAddrRet.stdout.decode("utf-8").split()[0]
-debugPrint("	got address " + addressStr)
-
-# generate blocks for funds
-genInitCmdArgs = initCmdArgs.copy()
-genInitCmdArgs.append('generate')
-genInitCmdArgs.append('101')
-genInitRet = subprocess.run(genInitCmdArgs, capture_output=True)
-exitWithMessageIfError(genInitRet.stderr, btcStartingChain, 'Error getting address')
-debugPrint("	generated Initial blocks (101)")
-
-# send Txs
-TxCmdArgs = initCmdArgs.copy()
-TxCmdArgs.append('sendtoaddress')
-TxCmdArgs.append(addressStr)
-TxCmdArgs.append('amount_placeholder') # amount
-genInitCmdArgs[5] = '1' # change command's number of generated blocks to 1
-
-for txNum in range(0, utxo_size):
-	TxCmdArgs[len(TxCmdArgs)-1] = str(TX_DEFAULT_SENT_AMOUNT)
-	sendTxRet = subprocess.run(TxCmdArgs, capture_output=True)
-	exitWithMessageIfError(sendTxRet.stderr, btcStartingChain, "Error sending Txs, failed on tx number " + str(txNum))
-	if ((txNum+1) % 100) == 0:
-		debugPrint('sent 100 Txs')
-	if ((txNum+1) % TX_NUMBER_MAX_IN_BLOCK) == 0 and txNum != 0:
-		genRet = subprocess.run(genInitCmdArgs, capture_output=True)
-		exitWithMessageIfError(genRet.stderr, btcStartingChain, "Error generating block number " + str(txNum / TX_NUMBER_MAX_IN_BLOCK))
-genLastRet = subprocess.run(genInitCmdArgs, capture_output=True)
-exitWithMessageIfError(genLastRet.stderr, btcStartingChain, "Error generating last block, number:" + str(txNum / TX_NUMBER_MAX_IN_BLOCK))
-debugPrint("	sent all Txs")
-print('created starting block')
+# bitcoindCmdArgs[1] = '-conf=' + confFilePath
+# btcStartingChain = subprocess.Popen(bitcoindCmdArgs, stdout=subprocess.DEVNULL, stderr=subprocess.STDOUT)
+# debugPrint("	started bitcoind of starting chain node")
+#
+# debugPrint("	waiting for client to finish setup...")
+# time.sleep(5)
+# initCmdArgs = cliCmdArgs.copy()
+# initCmdArgs[3] = '-rpcport=' + str(rpcport)
 
 
-input('copy files and press enter')
-btcStartingChain.terminate()
-####### move files to directories of nodes
-#	TODO: copy files from start blockchain node
 
 
+
+
+
+
+
+print('running nodes...')
 ####### run clients
 os.chdir(binPath)
-bc = []
+btcClients = []
 for node in range(0, num_clients):
 	nodeDir = nodesPath + "node" + str(node)
 	confDefault[7] = 'datadir=' + nodeDir
-	confDefault[8] = 'loadblock=' + nodeDir + 'regtest/blocks/blk00000.dat'
 	port = BASE_PORT_NUM + node
 	confRegtest[0] = 'port=' + str(port)
 	rpcport = BASE_RPC_PORT_NUM + node
@@ -237,40 +208,103 @@ for node in range(0, num_clients):
 		text_file.write('\n\n[regtest]\n')
 		text_file.write(contentRegTest)
 
-	bc.append(subprocess.Popen(bitcoindCmdArgs, stdout=subprocess.PIPE))
-	print("started bitcoind")
-	print("pid of client is " + str(bc[node].pid))
-	print('')
-	time.sleep(5)
+	btcClients.append(subprocess.Popen(bitcoindCmdArgs, stdout=subprocess.DEVNULL, stderr=subprocess.STDOUT))
+	debugPrint("	started bitcoind, pid: " + str(btcClients[node].pid))
+time.sleep(5)
+
+print('setting up miner node...' )
+initCmdArgs = cliCmdArgs.copy()
+initCmdArgs[3] = '-rpcport=' + str(BASE_RPC_PORT_NUM + 0) # node 0 will be the miner
+
+# generate blocks for funds
+genInitCmdArgs = initCmdArgs.copy()
+genInitCmdArgs.append('generate')
+genInitCmdArgs.append('101')
+genInitRet = subprocess.run(genInitCmdArgs, capture_output=True)
+exitWithMessageIfError(genInitRet.stderr, btcClients, 'Error getting address')
+debugPrint("	generated 101 Initial blocks")
+
+# get address for Txs
+GetAddressCmdArgs = initCmdArgs.copy()
+GetAddressCmdArgs.append('getnewaddress')
+getAddrRet = subprocess.run(GetAddressCmdArgs, capture_output=True)
+exitWithMessageIfError(getAddrRet.stderr, btcClients, 'Error getting address')
+addressStr = getAddrRet.stdout.decode("utf-8").split()[0]
+debugPrint("	got address " + addressStr)
+
+# send Txs
+TxCmdArgs = initCmdArgs.copy()
+TxCmdArgs.append('sendtoaddress')
+TxCmdArgs.append(addressStr)
+TxCmdArgs.append('amount_placeholder') # amount
+genInitCmdArgs[5] = '1' # change command's number of generated blocks to 1
+
+for txNum in range(0, utxo_size):
+	TxCmdArgs[len(TxCmdArgs)-1] = str(TX_DEFAULT_SENT_AMOUNT)
+	sendTxRet = subprocess.run(TxCmdArgs, capture_output=True)
+	exitWithMessageIfError(sendTxRet.stderr, btcClients, "Error sending Txs, failed on tx number " + str(txNum))
+	debugPrintNNewLine('.')
+	if ((txNum+1) % 100) == 0:
+		debugPrint('')
+		debugPrint('sent 100 Txs')
+	if ((txNum+1) % TX_NUMBER_MAX_IN_BLOCK) == 0 and txNum != 0:
+		genRet = subprocess.run(genInitCmdArgs, capture_output=True)
+		exitWithMessageIfError(genRet.stderr, btcClients, "Error generating block number " + str(txNum / TX_NUMBER_MAX_IN_BLOCK))
+genLastRet = subprocess.run(genInitCmdArgs, capture_output=True)
+exitWithMessageIfError(genLastRet.stderr, btcClients, "Error generating last block, number:" + str(txNum / TX_NUMBER_MAX_IN_BLOCK))
+bestBlockHashMiner = genLastRet.stdout.decode("utf-8").split()[0] # best block hash of miner node
+debugPrint('')
+debugPrint("	sent all Txs")
+print('finished generating initial chain')
+
+# make sure all node are synced getbestblockhash
+bestBlockCmdArgs = initCmdArgs.copy()
+bestBlockCmdArgs.append('getbestblockhash')
+allSynced = False
+while not allSynced:
+	allSynced = True
+	for node in range(1, num_clients):
+		rpcport = BASE_RPC_PORT_NUM + node
+		bestBlockCmdArgs[3] = '-rpcport=' + str(rpcport)
+		bestBlockRet = subprocess.run(bestBlockCmdArgs, capture_output=True)
+		exitWithMessageIfError(bestBlockRet.stderr, btcClients, 'Error getting address')
+		bestBlockHashThisNode = bestBlockRet.stdout.decode("utf-8").split()[0]
+		if bestBlockHashThisNode != bestBlockHashMiner:
+			allSynced = False
+			time.sleep(0.5) # wait a bit and start over
+			break
+debugPrint("	all node are synced")
 
 
-# generateCmdArgs = cliCmdArgs.copy()
-# generateCmdArgs.append('')  # cmd
-# generateCmdArgs.append('')  # amount
-# generateCmdArgs[4] = 'generate'
-# generateCmdArgs[5] = '1'
-# for node in range(0, num_clients):
-# 	if node != 0 :
-# 		continue
-# 	rpcport = BASE_RPC_PORT_NUM + node
-# 	generateCmdArgs[3] = '-rpcport=' + str(rpcport)
-# 	genInitRet = subprocess.run(generateCmdArgs, capture_output=True)
-# 	time.sleep(1)
-# 	print(genInitRet.stdout.decode("utf-8"))
-# 	print(" ")
-# 	print(genInitRet.stderr.decode("utf-8"))
-# 	print('cmd generate')
-# 	print(generateCmdArgs)
+# start script for timing
+try:
+	serverRet = subprocess.Popen(['python3.7', parentDirPath + 'server.py', num_clients], capture_output=True, timeout=)
+except ValueError:
+	print('Error timing block spread')
+	print(str(ValueError))
+	for node in range(0,num_clients):
+		btcClients[node].terminate()
 
+# generate block for timing
+generateCmdArgs = cliCmdArgs.copy()
+generateCmdArgs.append('generate')  # cmd
+generateCmdArgs.append('1')  # amount
+generateRet = subprocess.run(genInitCmdArgs, capture_output=True)
+exitWithMessageIfError(generateRet.stderr, btcClients, 'Error generating block for timing')
+print('generated block for timing')
+
+# wait until server is done timing (all nodes got block)
+while p.poll() is None:
+    time.sleep(0.5)
+
+with open(parentDirPath + 'time.txt') as f:
+	timeGot= f.read()
+debugPrint('	time got ' + str(timeGot) + ' seconds')
 
 
 input("Press Enter to terminate")
 for node in range(0,num_clients):
-	bc[node].terminate()
+	btcClients[node].terminate()
 print("terminated")
 
-# out, err = bc[0].communicate()
-# print(out.decode("utf-8"))
-# if err is not None:
-# 	print(err.decode("utf-8"))
